Continue implementing basic operations on or with ValueReflection
- implement operations with booleans and integers,
- change ValueInteger's internal type to be an int,
- remove booleanValue() and longValue() methods in favour of getRawValue() with appropriate cast -- this allows for more flexibility around Reflection without impeding other functionalities,
- clean up any classes that made use of the removed methods,
- expand the test for ValueReflection.

TODO:
- continue implementing operations on and with ValueReflection,
- consider implementing ValueFloat, ValueDouble and ValueLong instead of having simplified types for the cost of flexibility with Reflection API constructors.

+++ Clove/src/values/ValueRational.java
@@ -50,7 +50,7 @@ public class ValueRational extends ValueAbstract {
 	 */
 	private Value tryInt(double v) {
 		if ((v == Math.floor(v)) && !Double.isInfinite(v))
-		    return new ValueInteger((long) Math.floor(v));
+		    return new ValueInteger((int) Math.floor(v));
 		else
 			return new ValueRational(v);
 	}
